Creation and first cleaning.
Uploading the final version of my previous CV and cleaning it up to
transform it into the Master CV. The idea is to create a source code
with absolutely everything I can put into it. Then I can comment in an
out the things I need in each case and create a CV best suited for each
job offer.